If the refcount table is too small, enlarge it
The function `max_refcount_table_size` will return a size suitable for
refcounting the data blocks, but not the refcount blocks (recall that
the refcount table entries point to clusters which contain the reference
counts). It will also fail to take into consideration the fact that the
file may legitimately have holes in.

This patch fixes an infinite loop caused when re-allocating the refcount
table _to the same size it already is_ by guaranteeing that the table
will be doubled in size in this case.

This is related to #8

Signed-off-by: David Scott <dave.scott@unikernel.com>