Increment the refcount of all header clusters
Previously we were incrementing the refcount of individual clusters,
which fails for very large files (e.g. 1PiB) when the l1_size is bigger
than a cluster.

This patch increments the refcount of all header clusters, where a
header cluster is anything less than the `next_cluster` high water mark
when we've allocated the header.

With this patch `qemu-img check` of a 1PiB empty file passes.

Signed-off-by: David Scott <dave.scott@unikernel.com>