If lazy_refcounts is set, then don't update the refcounts
This patch creates a new mutable field `lazy_refcounts` which is
initially populated with the policy from the header. The value will
diverge from the header when we are doing things like repairing
the refcount table because we don't want to update the on-disk
header until we've finished.

We maintain the property that the in-memory copy of the header is
a faithful copy of the header on-disk.

Signed-off-by: David Scott <dave@recoil.org>