Optimise away a `flush` when allocating a data and L2 cluster
When the L1 entry is 0 i.e. undefined, we allocate 2 clusters: one
to contain the L2 cluster and the other for the data. We then issue
2 writes:

- L1[index] <- &L2
- L2[index'] <- &data

These writes can happen in either order, since if the L1 is persisted
then the L2 block is empty (which is valid); or if the L2 is persisted
then it is unreferenced from the L1 (which is safe).

We only need one call to `flush` at the end.

Signed-off-by: David Scott <dave@recoil.org>