Decouple `discard` from `compact`
Previously we would perform a synchronous discard every time we exceeded
the compaction threshold. This patch instead (re-)starts a timer on every
`discard` over the threshold so we only `compact` when the `discard`s have
stopped for at least 1 second. This hopefully lets an `fstrim` command in
the VM complete before we start compacting. Hopefully this means that we
avoid pointlessly shuffling blocks which we're about to `discard`.

Signed-off-by: David Scott <dave@recoil.org>