Create an abstract function `compact_s`
Previously the logic to figure out what operations to perform to
compact the file was entwined with the I/O code for making it
happen. This patch separates the logic / planning code from the
I/O to make the logic easier to test and the I/O code easier to
modify.

Signed-off-by: David Scott <dave@recoil.org>