read, write, discard: cancel an in-progress compact
Previously the `compact` would continue and block until it had
finished. This isn't ideal because the blocked I/O may cause the client
to get confused / try to cancel the requests. For example the Linux
block layer becomes unhappy after 30s of delay which could easily
happen with a big .qcow file or a slow disk.

This patch makes `read`, `write` and `discard` cancel any in-progress
compact and schedule another for the future.

Signed-off-by: David Scott <dave@recoil.org>