connect: call `check` to sanity-check the image
Previously a corrupt image may be `connect`-able but later calls to
`read` or `write` might fail. This patch makes `connect` call `check`
which will cause the `connect` to fail if the image is obviously
corrupt.

This will make `connect` slightly slower, but it will cause the metadata
blocks to be cached speeding up future I/O.

Signed-off-by: David Scott <dave@recoil.org>