compact: correct off-by-ones in the initial free set calculation
The interval (a, b) means a..b including both a and b. Therefore if
the interval representing the L1 table is

  (l1_table_start_cluster, l1_table_start_cluster + l1_table_clusters - 1)

Previously this would mark more blocks as used than necessary, which is
safe. In the bitmap code the bounds check raises an exception. This
patch fixes the bounds and avoids the exception.

Signed-off-by: David Scott <dave@recoil.org>