Hide cluster offset calculations
Previously we would write

  let cluster, within = Physical.to_cluster

to compute the cluster on disk and the byte offset within it of a
particular address. This patch changes the API to

  let cluster = Physical.cluster
  let n = Physical.within_cluster

where the cluster is assumed to be an array of Physical addresses
(i.e. this code is only used for the L1/L2 tables) and `n` is the
index of the array, not the byte offset (where `8*n=offset`)

Signed-off-by: David Scott <dave@recoil.org>