Don't add clusters beyond the file end to the free list
The bitmap storing which clusters are free expands on demand and is not
shrunk. Therefore the file can become smaller than the bitmap. We should
not consider these blocks to be in the free list because blocks in the
free list contain junk and must be overwritten, unlike blocks created
by extending the file.

Signed-off-by: David Scott <dave@recoil.org>