allocate_clusters: return whether the clusters contain zeroes or not
When we extend the file to make a cluster, it will contain zeroes.

When we recycle an existing block, it will not contain zeroes and we
need to erase it.

Note this does not optimise out the zeroing in the case where real data
is about to be written to the cluster. This is tricky to do safely since
the data has to be written first, before the pointer.

Signed-off-by: David Scott <dave@recoil.org>