Call `flush` through `Recycler.flush`
This will allow careful management of state across flushes. For example
only blocks which have been zeroed before the flush is issued should be
added to the available list. Previously the flush could permute with an
erase and we would erroneously believe that newly erased blocks had been
flushed when they haven't.

Signed-off-by: David Scott <dave@recoil.org>