Hold a metadata mutex while walking or modifying L1/L2
Previously we held a mutex only while allocating, to prevent 2
threads from allocating different clusters for the same virtual address.

This patch holds a single mutex while reading or writing to the L1 and
L2 tables. This is held over

- read
- write
- discard
- reparent (part of a compact)

Note that the L1 and L2 tables are all cached in memory. If this is
too slow, we could add lazy write-back to this cache.

Signed-off-by: David Scott <dave@recoil.org>