Flesh out more of the background thread logic
When more junk data is available (<-- this needs to be when anything
happens to any of the counters), first check if we need to erase more.
If we don't need to erase more, consider whether we should run a compact.

This doesn't address the need to call flush if the user doesn't.

Signed-off-by: David Scott <dave@recoil.org>