debug: improve the no leaked block assert
This now checks that sets of blocks which should have no intersection
actually do have no intersection. For example a block cannot be both
'junk' and 'erased' or 'available' and in 'refs'. The only exception
is the 'roots' set to which blocks are added while references are being
also added elsewhere.

Signed-off-by: David Scott <dave@recoil.org>