WIP: remove the t.next_cluster pointer
Since resizing the file is actually quite expensive, this patch resizes
in increments of 32MiB. The cluster map determines whether clusters are
free or not without having to constantly query/change the file size.

This patch also removes the cache of the `Mirage_block.info` since this
changes dynamically and should be queried before use.

Signed-off-by: David Scott <dave@recoil.org>