read: coalesce adjacent IOs for efficient use of readv
Previously we would run every cluster read concurrently, which is fine
if the overhead of a request is low. On Unix each request becomes a
syscall and the overheads quickly dominate for 64KiB default cluster
sizes. This patch scans over the list of IO operations and coalesces
adjacent ones to reduce the number of syscalls.

Signed-off-by: David Scott <dave@recoil.org>