compact: ensure we run enough passes
Previously we would run a single pass which wouldn't be quite enough
if we need to move an L2 metadata block which is written to in the
first pass. We run 3 passes: 2 should be enough, the extra pass is
there only as a sanity check.

Also specifically resize the file to the minimum size. While the file is
live we keep slack in the file, but offline we might as well reclaim all
the slack back.

Signed-off-by: David Scott <dave@recoil.org>