Avoid space on the end of the file getting into the `junk` list
We keep the invariant that free space on the end of the file (i.e.
after `get_last_block`) has been created by `ftruncate` and is already
erased and ready to allocate. Previously such space would end up on
the `junk` list too, which would let it become double allocated.
We fix this by deliberately shrinking the file on `connect` to avoid
recycling these blocks.

Signed-off-by: David Scott <dave@recoil.org>