Don't fail if a block is discarded after we start a compaction
Previously we triggered an assertion if we tried to move a block which
had been discarded after the compaction process has started. Instead
of failing, we should simply log a debug message. Ideally this won't
happen too often as we will try (in a later patch) to wait for the
discards to finish before triggering a compact.

Signed-off-by: David Scott <dave@recoil.org>