recycler: `get_moves` should omit in-progress moves
Clusters which are in state Copied are waiting for a `flush` to
transition to state `Flushed`. Previously `get_moves` would return these
same clusters as candidates for moving which would cause a recopy
and a state reset, preventing the move from being completed.

This patch causes `get_moves` to return the list of moves which
still need to be executed, since once a move has started it either
completes on a flush or is cancelled by a write.

Signed-off-by: David Scott <dave@recoil.org>