make_cluster_map should compute the header space properly
When qemu-img creates a .qcow2 it typically leaves a cluster free
in the middle of the header. This would cause `min_elt` to point to
this free cluster which could cause us to truncate the file too far
and erase the L1 table.

This patch assumes the first free block is after the header structures
i.e. after the refcount table and the L1 table. If the file is full
then the first_movable_cluster = the file size in clusters i.e. the
next block that will be allocated.

Furthermore this patch ensures that `get_last_block` will return
one less than the first free block if there are no bindings in the
map.

Hopefully this fixes some of the unit tests checking qemu-img compat.

Signed-off-by: David Scott <dave@recoil.org>