Make the cluster locks recursive
Previously the read/writer cluster locks were non-recursive, but this
was awkward because we wanted to take a batch of reads/writes, lock
all the metadata clusters for reading, acquire a write lock too if
necessary where some of the metadata clusters are shared.

This patch makes the locks optionally recursive: a `client` owns a set
of locks and can acquire the same lock it already has.

This patch also includes a test case which verifies that a compact
in the middle of a `write`, (a) doesn't deadlock; and (b) doesn't lose
data.

Signed-off-by: David Scott <dave@recoil.org>