Don't hold the global lock while updating references
In dd2bcabbefbd793ebd7362d5254874da26e6dd5c we acquired the global lock
in the `update_references` function to prevent racing with an explicit
compact triggered by the user. We now have automatic compact and the
rule that only one compact should be triggered at a time.

This should hopefully prevent starving threads in `read`/`write` during
a compact.

Signed-off-by: David Scott <dave@recoil.org>