On moving an L2 cluster, update the back reference map
The back reference map (`t.Qcow_cluster_map.refs`) tracks the cluster
which points to this one, and which must be updated when this one moves.
In the case where we're moving an L2 cluster, all the back references
which still point to the old L2 cluster need to be updated to point
to the new L2 cluster.

Signed-off-by: David Scott <dave@recoil.org>