Hold a read lock on the L1 table entry
During a read or write operation, we look up the L1, then the L2,
then finally the data cluster. Sometimes we will need to allocate
the data cluster and write to the L2. Holding the read
lock on the L1 prevents a background compact from moving the L2
cluster while we're about to write the new data reference to it.

This patch causes read and write to maintain read locks on the
required L1 and L2 clusters during the operation.

Signed-off-by: David Scott <dave@recoil.org>