Check that roots are disjoint
We maintain a set of clusters as 'global roots' while they aren't
yet registered in an L1 or L2 metadata block. Since we store these
roots as a set, we must not allow any overlap between elements
otherwise we will erroneously deregister a root early. This patch
verifies that the intersection between the new roots and the existing
roots is empty.

Signed-off-by: David Scott <dave@recoil.org>