Register moves, junk before copying blocks
Previously we would register moves and remove from junk as we copied
each cluster. This meant that a destination cluster would remain in
the junk list for a while after the move was computed, and could be
used by the erasure thread. This patch registers the moves immediately
and removes the destinations from junk immediately to prevent that
from happening.

Signed-off-by: David Scott <dave@recoil.org>