Hide the low-level *.remove functions
Previously the Qcow_cluster_map users would call `add` and `remove`
directly. This patch removes the `remove` case, making the code
less likely to lose track of clusters altogether. To make this work
the `resize` function is modified to explicitly assume that clusters
created from nothing on the end of the file don't need to be zeroed
and are added to the `Available` set.

Signed-off-by: David Scott <dave@recoil.org>