Simplify dump of lock state
The lock datastructure can be serialised with either the cluster as
the focus (answering the question 'which clients have locked this')
or with the client as the focus (answering the question `which locks
does this client own`). Previously we printed with a cluster focus
but it's easier to relate to the intended I/O operation if we take
a client focus. This patch changes to a client focus i.e. if a
write operation fails it will print the locks owned by the write
client. This should allow us to see whether (for example) we have
read locks on the metadata clusters or not.

Signed-off-by: David Scott <dave@recoil.org>