When copying a block, check for move cancellation
Previously we would schedule the move of a block which could be
discarded in the meantime. We would then attempt to read from the
cluster only to find that it isn't there.

This patch cancels any move of a cluster when the cluster is unreferenced
(and added to Junk). This patch also holds the write lock on a cluster
while it is being discarded: previously we would only hold the write
lock on the cluster containing the reference to this one. This
ensures that an in-progress copy cannot interleave with a discard which
truncates the file while the copy is still running.

Signed-off-by: David Scott <dave@recoil.org>