get_last_block: include the source of moves
Previously we could get a move to the `Referenced` stage at which point

- the move is still in-progress: it's in the map
- `get_last_block` would not include the source block

This could lead to the file being truncated before the reference change
has been flushed. This patch includes the moves in the `get_last_block`
calculation. Note the destination blocks were already included in the
`Copies` set.

Signed-off-by: David Scott <dave@recoil.org>