debug: the `refs` set may overlap with `copies`
Previously we expected `refs` to be disjoint from `copies` but it's
important to keep a cluster in `copies` to ensure `get_last_block`
takes it into account until after the move is complete i.e. after
the reference has been updated and flushed. Therefore it's ok for
`refs` and `copies` to overlap.

Signed-off-by: David Scott <dave@recoil.org>