prometheus: use Counter for GC cluster transitions
Previously we used Gauges to record the current number of blocks in each
state at the sample time. However the sampling is infrequent and most of
the transitions are missed. This patch switches to cumulative Counters
for these states, so we can see "GC activity" e.g. total number of blocks
erased (rather than number of blocks which are in state 'erased' when the
sample is taken).

This patch leaves the `used` and `max_cluster` as Gauges.

Signed-off-by: David Scott <dave@recoil.org>