perf: make interval union calculations faster
The IntervalSet `union` is roughly:

  let union a b = fold add a b

where elements of `a` are added to `b`. Therefore if one of the
sets is much smaller than the other we should make it the 1st
argument.

Running a `qcow-tool compact` on files created with
`qcow-tool pattern --pattern 2` shows the following speedup:

Size/GiB      Before/seconds    After/seconds
1             20                1
2             87                3
8             1369              15
16            ??                25
64            ??                139

Signed-off-by: David Scott <dave@recoil.org>