cli: lock the file unless it's read-only
The read-only operation `check` is specifically designed to cope with
racing with an application writing to the disk by retrying if it detects
a reference to a block outside the file. We don't need to lock the disk
at all in this case.

All other operations require the disk to be locked.

Signed-off-by: David Scott <dave@recoil.org>