Use to_string instead of sexps for debugging
This removes the dependency to `sexplib` and `ppx_sexp_conv`.
It was only required to print error messages to the standard error
stream in the test suite.

Instead, this relies on a `to_string` function in the input signature.
This maps what can be found in `Int64`, and `Int` in recent versions on
the compiler.

As a bonus, a pretty printer is added to the output signature.